services/auth/requirements.txt
fastapi
uvicorn
asyncpg
python-jose
passlib[bcrypt]
pydantic
python-dotenv
structlog
prometheus-fastapi-instrumentator
pydantic-settings
pydantic[email]